Avoid fetching all remote branches by default
This works better for large, volatile repos with lots of remote
branches. Often I've found that the need to rebase against the
default branch is also separate from the need to pull the default
branch, which takes enough time to make it a separate command that
can still be used in combination with rebasing.